+++ tools/L10nImages/utils/Browser.js
