+++ tools/L10nImages/utils/Logger.js
