+++ tools/L10nImages/utils/SupportedLanguages.js
